rearrange code and directories
