+++ src/lib/_tests_/unit.test.ts
@@ -1,0 +1,8 @@
+import { expect } from 'chai'
+import option from '../../lib/options.js'
+
+describe('Option values', () => {
+  it('should return 9944', () => {
+    expect(option.API_PORT).to.equal(9944)
+  })
+})
